Step 4: use Apollo.mutate to vote for a post
We used 'Apollo.query' so let's repeat few steps.

It's a v0.0.1 of the tutorial sooo  I won't get into details about what mutation is, there are 1000000 of sources about that that.

'Apollo.mutation' gives an Observable and it's a cold one so we need to call subscribe to make mutation happend.

Now take a look at localhost:4200 (run yarn start if you didn't) and click on any vote button. There should be a http request, you can check it in dev tools. Buuut! Nothing updated, nothing changed, number of votes is still the same.

Why?

Because we used Apollo.query, which fetches data only once and completes the Observable.

Let's change it! Please, jump to the next step.