Step 5: Use Apollo.watchQuery to observe cache update
As you can see, we switched from .query to .watchQuery

Please open the app and click on vote button. See? Number of votes did change. That's because watchQuery method doesn't complete and is open for any new change, it keep listens to the Apollo's cache.

Quick info about how cache works. We fetch a list of posts, an array of them. Apollo Cache breaks an array and turns each element into a separate subject so every Post object is written to the cache and a list itself contains just a list of references (like pointers in c++, an address). It's important to know that by default Apollo handles on those objects that contains 'id' property.

If we make a mutation and receive the changed object as the result and that object contains an 'id' property then Apollo Cache knows that something changed in that object. So it updates the object and spreads the news to all watched queries.

Your query that contains a list of posts and that object, receives that news and your UI gets updated.

Isn't that cool?