Step 7: Install graphql code generator with apollo-angular template
We've got an app that uses GraphQL, is strongly typed, takes advantage of Apollo's cache and so on but the whole process took quite some time.

What if I told you that you don't have to do more that half of it?

There's a tool we created, it's called 'graphql-code-generator'. It allows to generate code based on GraphQL Schema. Since every type in GraphQL is strongly typed you can do amazing things thanks to that.

GraphQL Code Generator has many templates, we will use the one for apollo-angular.

Let's run:

$ yarn add -D graphql-code-generator

$ yarn add -D graphql-codegen-apollo-angular-template